+++ Backend/.env
@@ -1,0 +1,2 @@
+PORT=3000
+MONGO_URI=mongodb+srv://neerajadhikari037:neeraj015@cluster0.wvxpdij.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
