+++ Backend/model/User.model.js
@@ -1,0 +1,27 @@
+import mongoose from "mongoose";
+const userSchema = new mongoose.Schema({
+    name:{
+        type:String,
+        required:true,
+
+    },
+    email: {
+        type: String,
+        required: true,
+        unique: true
+    },
+    password: {
+        type: String,
+        required: true
+    },
+    phone: {
+        type: String,
+        required: true
+    }
+},{timestamps:true});
+
+
+export const User = mongoose.model('User', userSchema);
+
+
+
