+++ Frontend/index.html
@@ -63,32 +63,32 @@
             <div>
                 <img src="https://wallpapercave.com/wp/wp1987065.jpg" alt="">
                 <h1> Burger </h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
             <div>
                 <img src="https://traditionallymodernfood.com/wp-content/uploads/2021/08/south-indian-lunch-combo-scaled.jpeg" alt="">
                 <h1>Veg Thali</h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
             <div>
                 <img src="https://tse3.mm.bing.net/th?id=OIP.Xp2RQVjyCr3kN1G8VhcK_wHaE_&pid=Api&P=0&h=180" alt="">
                 <h1>Samosa</h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
             <div>
                 <img src="https://1.bp.blogspot.com/-A-3CXLzaZb4/Xyl_sR3c4LI/AAAAAAAAjKk/03Ysvlx6P0YPWD9WJVuv8dtDdDSjvjStQCLcBGAsYHQ/s3300/Grilled%2BBBQ%2BChicken%2B1.jpg" alt="">
                 <h1>Grilled BBQ Chicken</h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
             <div>
                 <img src="https://images.immediate.co.uk/production/volatile/sites/30/2020/12/Noodles-with-chilli-oil-eggs-6ec34e9.jpg" alt="">
                 <h1>Noodles</h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
             <div>
                 <img src="https://th.bing.com/th/id/R.0d1dfde60ab45c45ea89b7fc21bfb9fa?rik=2K%2f1t6hDdaEg2Q&riu=http%3a%2f%2fwww.pamperedchef.com%2ficeberg%2fcom%2frecipe%2f457000-lg.jpg&ehk=RwDn2Zxwuqyp8VeqZynlGRnvhZEHHsD488cdSHxJnQs%3d&risl=&pid=ImgRaw&r=0" alt="">
                 <h1>Dessert</h1>
-                <button>order now</button>
+                <button><a href="addcart.html">order Now</a></button>
             </div>
         </div>
