+++ Frontend/login.html
@@ -119,7 +119,7 @@
         };
 
         // Send data to the backend
-        fetch('http://localhost:5000/api/login', {
+        fetch('http://localhost:3000/api/login', {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json'
