+++ Frontend/register.html
@@ -98,7 +98,7 @@
                 };
 
                 // Send data to the backend
-                fetch('http://localhost:5000/api/register', {
+                fetch('http://localhost:3000/api/register', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json'
